requirements.txt
# python==3.9.12
vllm
modelscope
tiktoken
pdfplumber
PdfReader
PyPDF2
langchain
langchain-community
langchain-huggingface
jieba
rank_bm25
sentence-transformers
faiss-gpu
text2vec==1.2.9